+++ sdk/communication/azure-communication-chat/azure/communication/chat/_utils.py
@@ -15,11 +15,11 @@ def return_response(response, deserialized, _):  # pylint: disable=unused-argume
 
 class CommunicationErrorResponseConverter(object):
     """
-    Util to convert to List[Tuple[ChatThreadParticipant, Optional[AddChatParticipantsErrors]]
+    Util to convert to List[Tuple[ChatParticipant, Optional[AddChatParticipantsErrors]]
 
     This is a one-way converter for converting the follwing:
-     - AddChatParticipantsResult -> List[Tuple[ChatThreadParticipant, AddChatParticipantsErrors]
-     - CreateChatThreadResult -> List[Tuple[ChatThreadParticipant, AddChatParticipantsErrors]
+     - AddChatParticipantsResult -> List[Tuple[ChatParticipant, AddChatParticipantsErrors]
+     - CreateChatThreadResult -> List[Tuple[ChatParticipant, AddChatParticipantsErrors]
     """
 
     @classmethod
@@ -28,21 +28,21 @@ class CommunicationErrorResponseConverter(object):
         """
         Util function to convert AddChatParticipantsResult.
 
-        Function used to consolidate List[ChatThreadParticipant] and List[ChatError]
-        into a list of tuples of ChatThreadParticipant -> ChatError. In case of no error, empty
+        Function used to consolidate List[ChatParticipant] and List[ChatError]
+        into a list of tuples of ChatParticipant -> ChatError. In case of no error, empty
         list is returned
 
         :param participants: Request object for adding participants to thread
-        :type: participants: list(~azure.communication.chat.ChatThreadParticipant)
+        :type: participants: list(~azure.communication.chat.ChatParticipant)
         :param chat_errors: list of ChatError
         :type chat_errors: list[~azure.communication.chat.ChatError]
-        :return: A list of (ChatThreadParticipant, ChatError)
-        :rtype: list[(~azure.communication.chat.ChatThreadParticipant, ~azure.communication.chat.ChatError)]
+        :return: A list of (ChatParticipant, ChatError)
+        :rtype: list[(~azure.communication.chat.ChatParticipant, ~azure.communication.chat.ChatError)]
         """
         def create_dict(participants):
             # type: (...) -> Dict(str, ChatThreadParticipant)
             """
-            Create dictionary of id -> ChatThreadParticipant
+            Create dictionary of id -> ChatParticipant
             """
             result = {}
             for participant in participants: