+++ sdk/communication/azure-communication-chat/samples/chat_client_sample_async.py
@@ -64,14 +64,14 @@ class ChatClientSamplesAsync(object):
         # [START create_thread]
         from datetime import datetime
         from azure.communication.chat.aio import ChatClient, CommunicationTokenCredential
-        from azure.communication.chat import ChatThreadParticipant
+        from azure.communication.chat import ChatParticipant
 
         # set `endpoint` to an existing ACS endpoint
         chat_client = ChatClient(endpoint, CommunicationTokenCredential(token))
         async with chat_client:
 
             topic = "test topic"
-            participants = [ChatThreadParticipant(
+            participants = [ChatParticipant(
                 user=self.user,
                 display_name='name',
                 share_history_time=datetime.utcnow()